+++ src/main/java/flightassistant/controllers/MyOrderController.java
@@ -1,11 +1,14 @@
 package flightassistant.controllers;
 
+import flightassistant.domain.City;
 import flightassistant.domain.MyOrder;
 import flightassistant.repositories.MyOrderRepository;
 import flightassistant.service.MyOrderService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -28,7 +31,13 @@ public class MyOrderController {
 
     @GetMapping("{id}")
     public MyOrder getOne(@PathVariable("id") MyOrder myOrder) {
-        return myOrder;
+        if (myOrder != null) {
+            return myOrder;
+        } else {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, MyOrder.class.getSimpleName() + " not found"
+            );
+        }
     }
 
     @PostMapping
@@ -39,13 +48,25 @@ public class MyOrderController {
 
     @PutMapping("{id}")
     public MyOrder update(@PathVariable("id") MyOrder myOrderFromDb, @RequestBody MyOrder myOrder) {
-        BeanUtils.copyProperties(myOrder, myOrderFromDb, "id");
-        return myOrderRepository.save(myOrderFromDb);
+        try {
+            BeanUtils.copyProperties(myOrder, myOrderFromDb, "id");
+            return myOrderRepository.save(myOrderFromDb);
+        } catch (Exception e) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, MyOrder.class.getSimpleName() + " not found"
+            );
+        }
     }
 
     @DeleteMapping("{id}")
     public void delete(@PathVariable("id") MyOrder myOrder) {
-        myOrderRepository.delete(myOrder);
+        if (myOrder != null) {
+            myOrderRepository.delete(myOrder);
+        } else {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, MyOrder.class.getSimpleName() + " not found"
+            );
+        }
 
     }
