+++ src/main/java/flightassistant/controllers/PlaneController.java
@@ -4,7 +4,9 @@ import flightassistant.domain.Plane;
 import flightassistant.repositories.PlaneRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -25,7 +27,13 @@ public class PlaneController {
 
     @GetMapping("{id}")
     public Plane getOne(@PathVariable("id") Plane plane) {
-        return plane;
+        if (plane != null) {
+            return plane;
+        } else {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, Plane.class.getSimpleName() + " not found"
+            );
+        }
     }
 
     @PostMapping
@@ -35,13 +43,25 @@ public class PlaneController {
 
     @PutMapping("{id}")
     public Plane update(@PathVariable("id") Plane planeFromDb, @RequestBody Plane plane) {
-        BeanUtils.copyProperties(plane, planeFromDb, "id");
-        return planeRepository.save(planeFromDb);
+        try {
+            BeanUtils.copyProperties(plane, planeFromDb, "id");
+            return planeRepository.save(planeFromDb);
+        } catch (Exception e) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, Plane.class.getSimpleName() + " not found"
+            );
+        }
     }
 
     @DeleteMapping("{id}")
     public void delete(@PathVariable("id") Plane plane) {
-        planeRepository.delete(plane);
+        if (plane != null) {
+            planeRepository.delete(plane);
+        } else {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, Plane.class.getSimpleName() + " not found"
+            );
+        }
 
     }
